+++ test/support/sqlite_repo.ex
@@ -1,0 +1,5 @@
+defmodule AshPrefixedId.Test.SqliteRepo do
+  @moduledoc false
+
+  use AshSqlite.Repo, otp_app: :ash_prefixed_id
+end